requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
Pillow>=10.0
mercadopago>=2.4.0
psycopg2-binary>=2.9
dj-database-url>=2.1
whitenoise>=6.6
